fix: mf-4963 trusted tokens could include blocked tokens
